feat: Add annotation for preserving resources deletion (#18)
Added 'edp.epam.com/preserve-resources-on-deletion: true' annotation processing to prevent the operator from deleting resources from keycloak.

+++ controllers/keycloakrealmcomponent/terminator.go
@@ -10,21 +10,28 @@ import (
 )
 
 type terminator struct {
-	realmName     string
-	componentName string
-	kClient       keycloak.Client
+	realmName                   string
+	componentName               string
+	kClient                     keycloak.Client
+	preserveResourcesOnDeletion bool
 }
 
-func makeTerminator(realmName, componentName string, kClient keycloak.Client) *terminator {
+func makeTerminator(realmName, componentName string, kClient keycloak.Client, preserveResourcesOnDeletion bool) *terminator {
 	return &terminator{
-		realmName:     realmName,
-		componentName: componentName,
-		kClient:       kClient,
+		realmName:                   realmName,
+		componentName:               componentName,
+		kClient:                     kClient,
+		preserveResourcesOnDeletion: preserveResourcesOnDeletion,
 	}
 }
 
 func (t *terminator) DeleteResource(ctx context.Context) error {
 	log := ctrl.LoggerFrom(ctx)
+	if t.preserveResourcesOnDeletion {
+		log.Info("PreserveResourcesOnDeletion is enabled, skipping deletion.")
+		return nil
+	}
+
 	log.Info("Start deleting KeycloakRealmComponent")
 
 	if err := t.kClient.DeleteComponent(ctx, t.realmName, t.componentName); err != nil {
